issue56 moved driver setup to before hook
