+++ de.dashup.application/src/test/java/de.dashup.test/utils/CucumberHooks.java
@@ -3,15 +3,72 @@ package de.dashup.test.utils;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import de.dashup.test.steps.GeneralStepdefs;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
 
 public class CucumberHooks {
+    private WebDriver driver;
     @Before
-    public void doDriverSetup(){
-
+    public void doDriverSetup() throws IOException {
+        InputStream is = this.getClass().getResourceAsStream("../../../../my.properties");
+        Properties p = new Properties();
+        p.load(is);
+        String name = p.getProperty("test.browser");
+        if (name != null) {
+            switch (name) {
+                case "chrome":
+                    driver = this.setUpChromeDriver();
+                    break;
+                case "firefox":
+                    driver = this.setUpFirefoxDriver();
+                    break;
+            }
+        } else {
+            driver = this.setUpChromeDriver();
+        }
+        GeneralStepdefs.setDriver(driver);
     }
 
     @After
     public void doTearDown(){
         GeneralStepdefs.getDriver().quit();
+    }
+
+    private WebDriver setUpChromeDriver() {
+        final DesiredCapabilities desiredChromeCapabilities = DesiredCapabilities.chrome();
+        final ChromeOptions chromeOptions = new ChromeOptions();
+
+        if (DriverUtil.getPathToChromeBinary() != null) {
+            chromeOptions.setBinary(DriverUtil.getPathToChromeBinary());
+        }
+
+        desiredChromeCapabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
+        System.setProperty("webdriver.chrome.driver", DriverUtil.getDriverPath("chrome"));
+        WebDriver returningDriver = new ChromeDriver(desiredChromeCapabilities);
+        returningDriver.manage().window().maximize();
+        return returningDriver;
+    }
+
+    private WebDriver setUpFirefoxDriver() {
+        final DesiredCapabilities desiredFirefoxCapabilities = DesiredCapabilities.firefox();
+        final FirefoxOptions firefoxOptions = new FirefoxOptions();
+
+        if (DriverUtil.getPathToChromeBinary() != null) {
+            firefoxOptions.setBinary(DriverUtil.getPathToFirefoxBinary());
+        }
+
+        desiredFirefoxCapabilities.setCapability(FirefoxOptions.FIREFOX_OPTIONS, firefoxOptions);
+        System.setProperty("webdriver.gecko.driver", DriverUtil.getDriverPath("firefox"));
+        WebDriver returningDriver = new FirefoxDriver(desiredFirefoxCapabilities);
+        returningDriver.manage().window().maximize();
+        return returningDriver;
     }
 }
